# Use unique in `LoggerProvider` for storing loggers