--- a/sdk/log/provider_test.go
+++ b/sdk/log/provider_test.go
@@ -302,14 +302,14 @@ func TestLoggerProviderLogger(t *testing.T) {
 		p := NewLoggerProvider()
 
 		l0, l1, l2 := p.Logger("l0"), p.Logger("l1"), p.Logger("l0", log.WithInstrumentationAttributes(attribute.String("foo", "bar")))
-		assert.NotSame(t, l0, l1)
-		assert.NotSame(t, l0, l2)
-		assert.NotSame(t, l1, l2)
+		assert.NotEqual(t, l0, l1)
+		assert.NotEqual(t, l0, l2)
+		assert.NotEqual(t, l1, l2)
 
 		l3, l4, l5 := p.Logger("l0"), p.Logger("l1"), p.Logger("l0", log.WithInstrumentationAttributes(attribute.String("foo", "bar")))
-		assert.Same(t, l0, l3)
-		assert.Same(t, l1, l4)
-		assert.Same(t, l2, l5)
+		assert.Equal(t, l0, l3)
+		assert.Equal(t, l1, l4)
+		assert.Equal(t, l2, l5)
 	})
 }