Replace internal map for keeping track of loggers in `LoggerProvider` with `unique` feature.