@pellared 2025-03-12
> there is not much doc tbh.

True. You can also check: https://victoriametrics.com/blog/go-unique-package-intern-string/

> Applying to Scope will only help comparing the values (still that would be acceptable to add).

TBH, the more I think about it I am not sure it will help at all. It is something that has to be measured (benchmarks).