@yago-123 2025-03-13
Closing this, apparently it only helps with amount of bytes allocated.

Old version vs. this branch version: 
```
benchmark                                                                old ns/op     new ns/op     delta
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     778           1903          +144.44%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     642           2757          +329.71%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     767           3790          +394.20%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     660           4317          +554.19%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     654           4951          +657.38%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            784           794           +1.17%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            785           965           +22.92%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            782           928           +18.65%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            779           1024          +31.45%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            819           1023          +24.86%

benchmark                                                                old allocs     new allocs     delta
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              6              +100.00%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              6              +100.00%

benchmark                                                                old bytes     new bytes     delta
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     420           220           -47.62%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     326           221           -32.21%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     403           224           -44.42%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     337           231           -31.45%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     338           233           -31.07%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            377           227           -39.79%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            380           231           -39.21%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            380           231           -39.21%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            380           234           -38.42%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            394           229           -41.88%
```


Old version vs. `map[unique.Handle[instrumentation.Scope]]` key (in this case not even a single metric improves): 
```
benchmark                                                                old ns/op     new ns/op     delta
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     778           1293          +66.09%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     642           1208          +88.28%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     767           1250          +62.99%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     660           1298          +96.70%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     654           1285          +96.57%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            784           1532          +95.33%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            785           1599          +103.72%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            782           1652          +111.23%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            779           1687          +116.56%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            819           1721          +110.06%

benchmark                                                                old allocs     new allocs     delta
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              7              +133.33%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            3              7              +133.33%

benchmark                                                                old bytes     new bytes     delta
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     420           437           +4.05%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     326           417           +27.91%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     403           423           +4.96%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     337           427           +26.71%
BenchmarkLoggerRetrieval/Create_different_loggers_single-threaded-12     338           406           +20.12%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            377           422           +11.94%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            380           421           +10.79%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            380           424           +11.58%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            380           424           +11.58%
BenchmarkLoggerRetrieval/Create_different_loggers_parallel-12            394           425           +7.87%

```