@dmathieu 2025-03-13
Reducing the number of allocations is a pretty good thing.