@yago-123 2025-03-13
For the shake of science. Added prepopulated (I hope the impl. is correct) values old vs. new: 
``` 
benchmark                                                                                                       old ns/op     new ns/op     delta
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values-12           827           715           -13.49%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values-12                  759           890           +17.33%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values#01-12        27.7          5619          +20170.56%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values#01-12               64.0          932           +1356.79%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values-12        779           1017          +30.52%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values-12               739           950           +28.64%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values#01-12     35.6          5928          +16570.42%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values#01-12            68.8          937           +1260.35%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values-12        761           945           +24.30%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values-12               722           1115          +54.33%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values#01-12     35.8          6173          +17143.02%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values#01-12            68.6          1062          +1447.20%

benchmark                                                                                                       old allocs     new allocs     delta
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values-12           2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values-12                  2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values#01-12        0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values#01-12               0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values-12        2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values-12               2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values#01-12     0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values#01-12            0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values-12        2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values-12               2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values#01-12     0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values#01-12            0              4              +Inf%

benchmark                                                                                                       old bytes     new bytes     delta
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values-12           379           217           -42.74%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values-12                  368           220           -40.22%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values#01-12        0             193           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values#01-12               0             189           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values-12        373           218           -41.55%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values-12               356           223           -37.36%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values#01-12     0             200           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values#01-12            0             192           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values-12        369           217           -41.19%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values-12               347           212           -38.90%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values#01-12     0             198           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values#01-12            0             187           +Inf%

```