@yago-123 2025-03-13
For the shake of science. Added the benchmarks for prepopulating the loggers with values and added bench for asking all the times the same different loggers in each iteration. A bit strange results: 
```
benchmark                                                                                                         old ns/op     new ns/op     delta
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_0_prepopulated_values-12        811           757           -6.61%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_1000_prepopulated_values-12     740           854           +15.36%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_5000_prepopulated_values-12     757           919           +21.46%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_0_prepopulated_values-12               792           872           +10.06%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_1000_prepopulated_values-12            889           939           +5.55%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_5000_prepopulated_values-12            766           925           +20.70%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values-12             28.1          5694          +20185.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values-12          35.8          5844          +16237.71%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values-12          35.9          5694          +15782.85%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values-12                    63.0          879           +1294.54%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values-12                 70.0          883           +1161.72%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values-12                 69.0          900           +1204.03%

benchmark                                                                                                         old allocs     new allocs     delta
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_0_prepopulated_values-12        2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_1000_prepopulated_values-12     2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_5000_prepopulated_values-12     2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_0_prepopulated_values-12               2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_1000_prepopulated_values-12            2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_5000_prepopulated_values-12            2              5              +150.00%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values-12             0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values-12          0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values-12          0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values-12                    0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values-12                 0              4              +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values-12                 0              4              +Inf%

benchmark                                                                                                         old bytes     new bytes     delta
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_0_prepopulated_values-12        387           214           -44.70%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_1000_prepopulated_values-12     354           218           -38.42%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_single-threaded_with_5000_prepopulated_values-12     354           218           -38.42%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_0_prepopulated_values-12               361           223           -38.23%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_1000_prepopulated_values-12            374           222           -40.64%
BenchmarkLoggerRetrieval/Retrieve_different_logger_each_time_parallel_with_5000_prepopulated_values-12            361           220           -39.06%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_0_prepopulated_values-12             0             197           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_1000_prepopulated_values-12          0             200           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_single-threaded_with_5000_prepopulated_values-12          0             195           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_0_prepopulated_values-12                    0             191           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_1000_prepopulated_values-12                 0             200           +Inf%
BenchmarkLoggerRetrieval/Retrieve_same_logger_each_time_parallel_with_5000_prepopulated_values-12                 0             198           +Inf%
```